Add certificate management support
Certificates for payload traffic is provided through Barbican
secrets store and must be uploaded by the end-user.

Certificates for authentication between Octavia controllers
and its Amphorae instances are issued locally on the Octavia
controller.

At the time of this writing this is the only supported
alternative upstream after the retirement of the Anchor
project [0].

Note that the locally generated certificates are not used for
any load balancer payload data.

0: https://review.openstack.org/#/c/597022/

Change-Id: I38c3e39de4d4cb173970a615f270839871d13910