Add action for configuring mgmt network for comms with lb instances
End user is expected to run the `configure-resources` action
after the deployment is complete and the cloud APIs are
reliably available.

The end user may rely on our defaults or may create the
required resources themselves.  The same action is used to
prompt configuration of the Octavia service and whether it
does creation or just discovery depends on the setting of
`create-mgmt-network` and `custom-amp-flavor-id` configuration
options.

Switch to using the more recent ``section-database`` part
for database configuration.

Change-Id: I9529a8a633ef0ba696c22570ec388991ba408ac4