Use system CA Certificate bundle when configuring resources
The charm bundles its dependencies to avoid the boundary
violation of the charm being dependent on the software it
manages.

As a side effect the charm will not use the distro packaged
version of ``python3-certifi`` which contains a patch [0] to
make consumers load the system wide CA Certificate bundle.

As a workaround we explicitly point our OpenStack client
operations to the system wide bundle.

0: https://git.launchpad.net/ubuntu/+source/python-certifi/tree/debian/patches/0001-Use-Debian-provided-etc-ssl-certs-ca-certificates.cr.patch

Change-Id: Iad466c7ff6cf680f74168852afea4a67815d0249
Closes-Bug: #1819205