Backport Zaza reconfiguration
By splitting up Zaza from the openstack-charms specific tests,
we can more easily guarantee test stability, while not harming
out ability to improve on the testing library during a cycle.

This change backports the zaza namespace change landing, and
adds a stable branch of the zaza-openstack-charms project to
the test-dependencies.

Change-Id: Ib641db1bbddedd86d85ba096e45032e1a7dc641c