Enable API TLS in the functional test
Add vault to the functional test bundle and enable TLS on all
charms that support it.

We want the functional tests to validate operation of Octavia with
multiple units.  When using TLS this will not work without also
enabling HA because of certificate hostname mismatch.  Remove all
non-HA bundles and do HA by default.

Change-Id: Ib1aafe0c26abc8876e073942fda2e4a1ecabbbb2