Add support for setting amphora ssh key
Allows a public key to be added to Amphora instances
to allow ssh access for e.g. debug purposes.

Without the ability to have amphora vms allow ssh access,
it can be impossible to debug issues that occur inside
the vm since it is not possible to modify an existing
image or vm without the vm being recreated. Adding
support for providing a key allows admins to have the
option to access amphora vms should something go
wrong. The default remains to not provide access.

Change-Id: I1d91741eeec9b20b80bbb907a8aeb9166831acbe
Closes-Bug: #1843625