Add Octavia OVN provider driver support
For use cases that do not require TLS termination or load balancing
based on higher layer characteristics, there is a Octavia OVN provider
driver that delegates load balancing to the network itself and does
not require VMs to handle the traffic.

Use the common ``parts/section-keystone-authtoken`` for keystone
auth in templates.

Change-Id: I1b9704c9e040eef821cef1e4f16faa0b18dce85e