Use separate application for Octavia's ovn-chassis subordinate
Just like with legacy Neutron ML2+OVS, the recommended deployment
topology for Octavia is to use a separate application for the
ovn-chassis subordinate co-located with Octavia.

We do this because the ovn-chassis application related to
nova-compute units will most likely have configuration set that
makes no sense for the Octavia units. (For example physical port
configuration, DPDK, SR-IOV, Hardware offloading etc etc)

Not adding this in the original Octavia OVN gate bundles was an
oversight, let's get it in.

Change-Id: I5852b3298d3d3e26ad077261fea07eb1d99d6e1c