Increase available cores and memory for octavia unit
The charm Octavia gate job has started to OOM for the Octavia
units as detailed on the bug.

This change addresses both that issue and attempts to increase the
velocity of the test by adding CPU cores which might help for the
time it takes to compile binary dependencies on initial deploy
and during the policyd tests.

Note that instead of explicitly mentioning amount of memory we
assume a flavor with more cores to also have more memory.

Func-Test-Pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/418
Change-Id: I0f332c5f6d97ed99f59c9985ff4a033920708bf2
Closes-Bug: #1895319