Gate reactive handlers on update-status
This patch gates many of the 'active' handlers when the update-status
hook is running.  This is to prevent the charm from making changes to
services, re-rendering configs, or updating relations (or attempting to
act on them) during an update-status hook.  This also speeds up the
update status so that it only does checking.  The related bug is an
example of what can happen if handlers are not masked during
update-status.

Change-Id: Ibabc74f9cd8c0130bcea76c455c179af31bc96a7
Related-Bug: #1895628