Specify OVN driver_agent for Victoria+
In Ussuri the ovn driver agent appears to have been included in
the wsgi app server. In victoria it has been moved out and needs
to be specified as a driver agent so it is started correctly.

Please note that although the docs refer to an amphora agent
it does not seem to exist and is not listed in octavia
entry points *1

*1 https://opendev.org/openstack/octavia/src/branch/stable/victoria/setup.cfg#L73

Closes-Bug: #1915830
Change-Id: I09dc3527bd19997ecd68ebe58acb4ddc727d9759