Ensure octavia-driver-agents gets installed
Due to a 'quirk' being fixed, the octavia-driver-agents don't get
installed if the ovsdb-subordinate  relation is the last to be made.
This is due to the update-status hook gating the reactive handler from
running which 'fixed' the issue.  Looking deeper, if the handler isn't
the first to run, then the charm's packages won't get initialised
properly.

This patch fixes that by using the dynamic properties 'all_packages' and
'full_service_list' available in charms.openstack core for the Charm
class.  This means that it doesn't matter when the flag is set, as long
as it is before the property is accessed.  Looking at the handler, this
will be in the right place for the install.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/514

Change-Id: I5fd75c9d371390bca402d6a3a264421a44fd092a
Closes-Bug: #1916764