Enable OVN L4 provider from Victoria onward
The driver is unreliable at Ussuri.

Also ensure we restart the ovn-driver-agent when configuration is
written to disk.

Ensure Port Security is enabled as this is a requirement for
reliable operation of the Amphora driver also exercised by the
functional tests.

Enable focal/groovy-victoria gates and re-enable Ussuri OVS gate.

Closes-Bug: #1927960
Closes-Bug: #1927962
Related-Bug: #1927959
Change-Id: I1bafd9eecb83aefb425bc5286f23d47c0885ac8e