Make Amphora provider driver optional
Octavia supports multiple provider drivers. The reference Amphora
provider driver is distributed as part of the Octavia software,
and is enabled by default.

Add a configuration option which allows to not enable the Amphora
provider driver. This is useful for use cases where a different
driver is desired, and removes the checks for Amphora related
mandatory configuration options.

Change-Id: I4d14056bf901d813ac53fde9ccd4e8bdb56e3c55