Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I50e457e3b9affdd0ca83602c1e580b75c6aa5434
(cherry picked from commit e3b8b03378793562740590d8f849876aee0d1f15)